Move loan change request specs into their own files.
